refactor ProfilePage
